UserAddress

+++ GoldMobile/src/com/csj/gold/mobile/vo/MobileParams.java
@@ -13,7 +13,7 @@ public class MobileParams {
 	
 	private String dateSources;//数据来源
 	private String userSources;//客户来源
-	private Integer userId;//用户ID
+	private Long userId;//用户ID
 	private String clientType;
 	private String businessSources;//业务来源
 	private String clientVersion;//
@@ -36,11 +36,11 @@ public class MobileParams {
 		this.userSources = userSources;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
